tlshd: Fix an implicit type conversion
implicit conversion loses integer precision: 'long' to 'int'

As far as I can tell, openat2() returns a zero or -1, so this
conversion is harmless. Annotate it.

Reported-by: Parfait 10.2 (#2028)

+++ src/tlshd/config.c
@@ -113,7 +113,7 @@ static int tlshd_file_open(const char *pathname)
 		.resolve	= RESOLVE_NO_SYMLINKS,
 	};
 
-	return syscall(SYS_openat2, 0, pathname, &how, sizeof(how));
+	return (int)syscall(SYS_openat2, 0, pathname, &how, sizeof(how));
 }
 #else
 static int tlshd_file_open(const char *pathname)
